requirements.txt
streamlit
pandas
openai
pdfplumber
python-dotenv
reportlab
plotly
pydantic